adds printing
